LPS-88715 Extract MediaObjectImpl to its own class
